requirements.txt
huggingface-hub==0.29.2
imageio==2.37.0
ipykernel==6.29.5
ipython==8.18.1
matplotlib==3.9.4
mmcv==1.3.1
numpy==1.25.2
openai==1.3.4
opencv-python==4.8.0.74
opendatalab==0.0.10
openxlab==0.0.11
packaging==24.2
pandas==2.1.2
Pillow==10.0.1
scikit-image==0.24.0
scipy==1.13.1
timm==1.0.15
tokenizers==0.13.3
tqdm==4.66.4
transformers==4.33.1


